+++ tests/net/auto_sntp/src/main.c
@@ -26,7 +26,7 @@
 K_SEM_DEFINE(l4_up, 0, 1);
 K_SEM_DEFINE(time_ref_updated, 0, 1);
 
-static void l4_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void l4_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 			     struct net_if *iface)
 {
 	ARG_UNUSED(iface);
