+++ tests/net/coap/general/src/main.c
@@ -37,7 +37,7 @@ struct cb_ctx {
 	uint32_t cb_count;
 };
 
-static void l4_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+static void l4_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 			     struct net_if *iface)
 {
 	ARG_UNUSED(iface);
